DAS-2183 - Implement variable.py.
